backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.66.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.14
langchain-core==0.2.34
langchain-google-genai==1.0.10
langchain-text-splitters==0.2.2
langsmith==0.1.104
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
orjson==3.10.7
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.32
tenacity==8.5.0
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.4
yarl==1.9.4
